income/expense requests now update net_worth

+++ server/src/routes/income.py
@@ -5,6 +5,7 @@ from pydantic.error_wrappers import ValidationError
 from pymongo.message import update
 from ..models.income import Income, UpdateIncomeModel, AddIncome
 from ..database.database import income_collection
+from ..database.database import net_worth_collection
 from ..utils.currency import get_exchange_rate_to_cad
 
 router = APIRouter()
@@ -22,6 +23,12 @@ def get_income_by_id(id):
 
 @router.post("/income/", response_description="Add new income", response_model=Income)
 def create_income(income: AddIncome = Body(...)):
+    if (net_worth_to_update := net_worth_collection.find_one()) is None:
+        raise HTTPException(status_code=404, detail=f"Net worths not found")
+
+    updated_current = net_worth_to_update["current"] + income.amount
+    updated_all_time_income = net_worth_to_update["all_time_income"] + income.amount
+
     if income.currency.lower() == "cad":
         income.exchange_rate = 1
     else:
@@ -37,6 +44,14 @@ def create_income(income: AddIncome = Body(...)):
             content=jsonable_encoder({"detail": exc.errors()}),
         )
 
+    net_worth_collection.update_one(
+        {"_id": net_worth_to_update["_id"]},
+        {
+            "$set": jsonable_encoder(
+                {"current": updated_current, "all_time_income": updated_all_time_income}
+            )
+        },
+    )
     insert_income = jsonable_encoder(insert_income)
     new_income = income_collection.insert_one(insert_income)
     created_income = income_collection.find_one({"_id": new_income.inserted_id})
@@ -47,9 +62,30 @@ def create_income(income: AddIncome = Body(...)):
     "/income/{id}", response_description="Delete income by id", response_model=Income
 )
 def delete_income_by_id(id):
+    if (income_to_delete := income_collection.find_one({"_id": id})) is None:
+        raise HTTPException(status_code=404, detail=f"Expense with id {id} not found")
+    if (net_worth_to_update := net_worth_collection.find_one()) is None:
+        raise HTTPException(status_code=404, detail=f"Net worths not found")
+
+    updated_current = net_worth_to_update["current"] - income_to_delete["amount"]
+    updated_all_time_income = (
+        net_worth_to_update["all_time_income"] - income_to_delete["amount"]
+    )
+
     delete_result = income_collection.delete_one({"_id": id})
 
     if delete_result.deleted_count == 1:
+        net_worth_collection.update_one(
+            {"_id": net_worth_to_update["_id"]},
+            {
+                "$set": jsonable_encoder(
+                    {
+                        "current": updated_current,
+                        "all_time_income": updated_all_time_income,
+                    }
+                )
+            },
+        )
         return JSONResponse(
             status_code=status.HTTP_200_OK,
             content=f"Income with id {id} was successfully deleted",
@@ -64,6 +100,15 @@ def delete_income_by_id(id):
     response_model=Income,
 )
 def update_income(id, income: UpdateIncomeModel = Body(...)):
+    if (income_to_update := income_collection.find_one({"_id": id})) is None:
+        raise HTTPException(status_code=404, detail=f"Expense with id {id} not found")
+    if (net_worth_to_update := net_worth_collection.find_one()) is None:
+        raise HTTPException(status_code=404, detail=f"Net worths not found")
+
+    amount_change = income.amount - income_to_update["amount"]
+    updated_current = net_worth_to_update["current"] + amount_change
+    updated_all_time_income = net_worth_to_update["all_time_income"] + amount_change
+
     if income.currency is not None:
         if income.currency.lower() == "cad":
             income.exchange_rate = 1
@@ -84,6 +129,17 @@ def update_income(id, income: UpdateIncomeModel = Body(...)):
 
         if update_result.modified_count == 1:
             if (updated_income := income_collection.find_one({"_id": id})) is not None:
+                net_worth_collection.update_one(
+                    {"_id": net_worth_to_update["_id"]},
+                    {
+                        "$set": jsonable_encoder(
+                            {
+                                "current": updated_current,
+                                "all_time_income": updated_all_time_income,
+                            }
+                        )
+                    },
+                )
                 return updated_income
 
     if (existing_income := income_collection.find_one({"_id": id})) is not None:
